parser: make accept_perms a propper perms_t constructor
Currently accept_perms is a weird helper fn for State construction.
We can move it into perms_t and make it a perms_t constructor which
will be used on State construction.

This will help logically associate the code with perms_t. Don't move
the code into the header as its a fairly large fn.

Signed-off-by: John Johansen <john.johansen@canonical.com>