apparmor: add permissions dedup cache
Currently we keep the set of permissions per state. But they need to
exist as external table the extended permissions format. Instead
of dedupping after the state machine build. Do a dedup during the
build, this will have the advantage that we can then use the perms_t
pointer as part of the minimization index instead of the whole
perms_t struct.

Signed-off-by: John Johansen <john.johansen@canonical.com>