tests: regression: Update socketpair test for upstream v9 socket behavior
Under upstream (v9) socket mediation, sockets passed to another process
get their labeling updated (similar to files), instead of keeping
just the opening cred. This is the technically correct approach for
ambient capabilities, but is a change in behavior.

This will not surface under normal read and write of the socket as
the task's creds are used not the cached socket label. It will however
make a difference when the socket label is used as a proxy which is
primarily when the socket is used for the peer label check.

Update the socket pair tests so they can pass with upstream socket
labeling, and compatibility patches applied.

Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from commit 43113b8999c14245cf6fff88b43bf6977e8d526e)
Signed-off-by: John Johansen <john.johansen@canonical.com>