tests: regression: update socketpair tests to detect different situations
The change in socket behavior, reflected in the previous patch only
occurs in some kernels. Better detect and report what is being used.
This not only helps with debugging but also is needed to ensure the
tests are correct when used on different combinations of the kernel and
parser.

Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from commit 35f1cb3d87056699edfbefb604b80489c0e3e12b)
Signed-off-by: John Johansen <john.johansen@canonical.com>