parser: libapparmor_re: fix addition of implied m when computing perms
Unfortunately the fix for permission computation that was causing
inconsistent builds still had a couple bugs around applying the
implied m permission.

This caused the implied m to not be applied in every situation that it
should have been. Resulting in profiles causing a denial when the
permission should have been allowed by policy.

The bugs are around the application of the implied permission to the
case where the file being executed is not owned by the user. The not
owner implied m permission will only be applied IF
  the rule specifying ix does not cause the owner permission to be set
  &&
  priority of the not owner implied permission bit is lower than
  the accumulated owner owner ix.

The bugs fixed here are strictly fail closed, they never cause
additional permissions to be granted.

Fixes: e7755c1fe ("Merge parser: libapparmor_re: fix inconsistent build due to bad implied m")
22a4e69f7 parser: libapparmor_re: fix inconsistent build due to bad implied m

Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from commit 181932869d578fe98b73e7201785ee2c75bc9957)
Signed-off-by: John Johansen <john.johansen@canonical.com>