regression tests: update logic to support v9 af_unix on legacy
This cleans up the and does a little rework on support tests for
unix and network mediation support.

The current tests fail because they don't bind support for a feature
between the parser and the user. Eg.

On a kernel that only supports v9 af_unix and a parser that only
supports v7 af_unix encoding the tests, checking for any af_unix
support and whether the parser supports "unix," rules will fail
because the parser can not encode v9 af_unix support, even though both
the kernel and parser "supports" af_unix.

To fix this add a set of support routines, that can test if the
parser supports different version encoding or downgrades, and then
build on those to test for support of combinations. Eg.

  parser supports v9 && kernel supports v9

and then build on those for generic tests
  supports_unix_rules

Update the individual regression tests and while we are at it clean
up some of the output formating.

Note: the v8 network mask in the af_unix_v7 abi is necessary to work
      around https://gitlab.com/apparmor/apparmor/-/issues/548

Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from commit 1224e343f3769b931b2ec725ee79070ad092aeb0)
Signed-off-by: John Johansen <john.johansen@canonical.com>