tests: regressions: Fix socket pair for v7 semantics
v7 unix socket label sets the socket to the opening tasks label and
never changes it.

Unfortunately the update to add v9 support got the label that would go
on the socket for v7 wrong. Using the label the test is transitioning
to instead of the label of the test when it opens the socket.

Fixes: 43113b899 ("tests: regression: Update socketpair test for upstream v9 socket behavior")
Signed-off-by: John Johansen <john.johansen@canonical.com>
(cherry picked from commit 3c09e1bd00ec74c5eb1771d3763265891c64ef46)
Signed-off-by: John Johansen <john.johansen@canonical.com>