parser: fix parsing of unix addresses with alternations
In AppArmor policy, unix addresses must start with @, and when trying
to use alternations {.,.}, the parser would confuse that with an
invalid variable, rightfully so. The problem is that this was
supported in previous versions of the parser, so in order to not
regress behavior, we are accepting such strings.

Fixes: https://gitlab.com/apparmor/apparmor/-/issues/557
Signed-off-by: Georgia Garcia <georgia.garcia@canonical.com>