parser: deprecate usage of alternation preceeded by @
In the past, the parser allowed the usage of alternations preceeded by
@, which was interpreted as the character @ followed by the expansion
of the alternation. For example: @{foo,bar} becomes @foo and @bar.

This is also the format of invalid set variables in policy, so in
order to avoid confusion in the future, this patch is deprecating the
former usage.

Signed-off-by: Georgia Garcia <georgia.garcia@canonical.com>