policy: add a 4.0-ip abi that can be used for fine grained inet mediation
Ubuntu is shipping a patch that provides fine grained mediation of ip
and ipv6 (see documentation in man apparmor.d). Support for it was
added in AppArmor 4.0 but the abi file was not updated to provide
support.

Provide an abi file so users can make use of fine grained inet mediation.

Signed-off-by: John Johansen <john.johansen@canonical.com>