docs: Add information about the reasons complain mode can fail
Complain mode is poorly understood, because it doesn't actually allow
everything, and can not be used to effectively turn off AppArmor, or
find all reasons an application may fail when apparmor is enabled.
Document when and and why complain mode may not convert a failure to
something that is allowed.

And while we are at it replace the cases of lower case apparmor that
makes sense to replace. Not every use is replaced because the lower
case form is appropriate when refering to specific files, commands, or
man pages.

BugLink: https://gitlab.com/apparmor/apparmor/-/issues/559
Signed-off-by: John Johansen <john.johansen@canonical.com>