tests: regressions: fix allow_all
The allow_all tests profile fails on systems where the specified abi
is not present. S the test should not make assumptions about the
available abi.

The allow_all test is supposed to reflect allow alls capability to
enable all mediation. So instead of providing a custom abi file for
the allow_all test we can use the kernel abi.

Signed-off-by: John Johansen <john.johansen@canonical.com>