Allow conditional assignment to be done without a space
Currently conditional assignment requires a space for tokenization.
This patch allows conditional assignment without a space between
the variable name and the conditional assignment ?= token.

This does technically break what is allowed as an id in some path
expressions by treating the combination of }? as invalid in an
id unless escaped or quoted.

Signed-off-by: John Johansen <john.johansen@canonical.com>