Add the ability to do an overriding assignment operator
The standard assignment operator does not allow resetting/overwriting
a variables values. If the assignment is being done to a variable that
already exist the assignment will fail.

Failure on reassignment for policy is often the desired behavior as it
forces the variable to be consistent for all of policy but some time
exceptions are needed. The current behavior prevents a user from being
able to explicitly override a variables values for a specific profile
or situation.

Add a new assignment operator that allows overriding the the values
assigned to a variable, to support situations that require an explicit
exception.

Signed-off-by: John Johansen <john.johansen@canonical.com>