init: split init based scripts/fns from the parser
Having the init scripts/fns in the parser directory is mixing
functionality and is just messy. Split them out into their own
dir.

For now the binutil/aa-load is left where it is. While it could be
moved to the init dir for non-systemd systems, until requested
just leave it where it is.

Signed-off-by: John Johansen <john.johansen@canonical.com>