parser: emit permsv2=3 along non-empty perms table
This allows the kernel to load the u32 accept2 table and correctly
discover and use the owner bit.

This patch needs more work, so that the value is only emitted for file
PDB and only when permstable32 is used.

Fixes: https://gitlab.com/apparmor/apparmor/-/issues/570

Signed-off-by: Zygmunt Krynicki <zygmunt.krynicki@canonical.com>
(cherry picked from commit 067cfc9f121f473e616bf10488c5e119a5a5326a)
Signed-off-by: John Johansen <john.johansen@canonical.com>