convert ancient use of PERMITTING, REJECTING, and unconstrained
PERMITTING and REJECTING were used in ancient (pre 2005) apparmor logs
before the effort to standardize some terms with selinux logs.

The terms were mapped
  fron             to
PERMITTING    -> ALLOWED
REJECTING     -> DENIED
unconstrained -> unconfined

Since PERMITTING and REJECTING won't appear in any apparmor logs
for the last 20 years, and the library is already mapping PERMITTING,
and ALLOWED to the same value, and REJECTING and DENIED to the same
value, it makes sense to remove the last vestiges of these terms
from the python utilities.

Signed-off-by: John Johansen <john.johansen@canonical.com>